# backend/internal/dto/dashboard_dto.go
package dto

import "time"

// ─── Dashboard Stats ──────────────────────────────────────────────────────────

type DashboardStats struct {
	CasesToday           int64   `json:"cases_today"`
	AwaitingReview       int64   `json:"awaiting_review"`
	AvgProcessingMinutes float64 `json:"avg_processing_minutes"`
	AiConfidenceAvg      float64 `json:"ai_confidence_avg"`
	ActiveCrawls         int64   `json:"active_crawls"`
}

type CasesByDay struct {
	Date  time.Time `json:"date"`
	Count int64     `json:"count"`
}

type EscalationTrend struct {
	Date  time.Time `json:"date"`
	Count int64     `json:"count"`
}

type Analytics struct {
	AutoApprovalRate    float64            `json:"auto_approval_rate"`
	TopUniversities     []UniversityCount  `json:"top_universities"`
	CountryDistribution []CountryCount     `json:"country_distribution"`
}

type UniversityCount struct {
	Name  string `json:"name"`
	Count int64  `json:"count"`
}

type CountryCount struct {
	Country string `json:"country"`
	Count   int64  `json:"count"`
}